♻️ change static Squirrel methods to a new export called varmintWorkspaceManager
