+++ packages/varmint/src/index.ts
@@ -2,4 +2,4 @@ export type * from "./cache-mode"
 export * from "./ferret"
 export * from "./sanitize-filename"
 export * from "./squirrel"
-export * from "./varmint-filesystem-state"
+export * from "./varmint-workspace-manager"
