+++ packages/varmint/src/squirrel.ts
@@ -4,7 +4,7 @@ import { inspect } from "node:util"
 
 import type { CacheMode } from "./cache-mode"
 import { sanitizeFilename } from "./sanitize-filename"
-import { storage } from "./varmint-filesystem-state"
+import { varmintWorkspaceManager as mgr } from "./varmint-workspace-manager"
 
 export type AsyncFunc = (...args: any[]) => Promise<any>
 
@@ -27,8 +27,11 @@ export class Squirrel {
 		this.mode = mode
 		this.baseDir = baseDir
 		this.rootName = sanitizeFilename(this.baseDir)
-		if (storage.initialized && !storage.getItem(`root__${this.rootName}`)) {
-			storage.setItem(`root__${this.rootName}`, this.baseDir)
+		if (
+			mgr.storage.initialized &&
+			!mgr.storage.getItem(`root__${this.rootName}`)
+		) {
+			mgr.storage.setItem(`root__${this.rootName}`, this.baseDir)
 		}
 	}
 
@@ -125,8 +128,11 @@ export class Squirrel {
 			for: (unSafeSubKey: string) => {
 				if (this.mode !== `off`) {
 					this.filesTouched.set(listName, new Set())
-					if (storage.initialized && !storage.getItem(`list__${listName}`)) {
-						storage.setItem(`list__${listName}`, `true`)
+					if (
+						mgr.storage.initialized &&
+						!mgr.storage.getItem(`list__${listName}`)
+					) {
+						mgr.storage.setItem(`list__${listName}`, `true`)
 					}
 				}
 				return {
@@ -144,8 +150,11 @@ export class Squirrel {
 							this.filesTouched.get(key)?.add(subKey)
 							const fileName = `${listName}__${subKey}` as const
 							const fileNameTagged = `file__${fileName}` as const
-							if (storage.initialized && !storage.getItem(fileNameTagged)) {
-								storage.setItem(fileNameTagged, `true`)
+							if (
+								mgr.storage.initialized &&
+								!mgr.storage.getItem(fileNameTagged)
+							) {
+								mgr.storage.setItem(fileNameTagged, `true`)
 							}
 						}
 						switch (this.mode) {
@@ -191,118 +200,4 @@ export class Squirrel {
 			}
 		}
 	}
-
-	public static startGlobalTracking(): void {
-		if (storage.initialized) {
-			console.error(
-				`💥 called startGlobalTracking, but the global cache was already initialized`,
-			)
-			return
-		}
-		storage.initialize()
-	}
-	public static flushGlobal(): void {
-		if (!storage.initialized) {
-			console.error(
-				`💥 called flushGlobal, but the global cache wasn't initialized with startGlobalTracking`,
-			)
-			return
-		}
-		const dirContents = fs.readdirSync(storage.rootDir)
-		const realRoots = new Map<string, string>()
-		const roots: `root__${string}`[] = []
-		const lists: `list__${string}`[] = []
-		const files: `file__${string}`[] = []
-		for (const dirContent of dirContents) {
-			if (startsWith(`root__`, dirContent)) {
-				roots.push(dirContent)
-			} else if (startsWith(`list__`, dirContent)) {
-				lists.push(dirContent)
-			} else if (startsWith(`file__`, dirContent)) {
-				files.push(dirContent)
-			}
-		}
-		const tree: Map<string, Map<string, Set<string>>> = new Map()
-		for (const root of roots) {
-			const rootName = root.replace(`root__`, ``)
-			tree.set(rootName, new Map())
-			const rootPath = storage.getItem(root)
-			if (rootPath) {
-				realRoots.set(rootName, rootPath)
-			} else {
-				console.error(
-					`💥 Could not find folder ${rootPath} referenced in the global cache`,
-				)
-			}
-		}
-		for (const list of lists) {
-			const listPath = list.replace(`list__`, ``)
-			const [listRootName, listName] = listPath.split(`__`)
-			const listRoot = tree.get(listRootName)
-			if (listRoot) {
-				listRoot.set(listName, new Set())
-			} else {
-				console.error(
-					`💥 Could not find root ${listRootName} for list ${listName}`,
-				)
-			}
-		}
-		for (const file of files) {
-			const filePath = file.replace(`file__`, ``)
-			const [listRootName, listName, subKey] = filePath.split(`__`)
-			const listRoot = tree.get(listRootName)
-			if (listRoot) {
-				const list = listRoot.get(listName)
-				if (list) {
-					list.add(subKey)
-				} else {
-					console.error(
-						`💥 Could not find list ${listName} for file ${filePath}`,
-					)
-				}
-			} else {
-				console.error(
-					`💥 Could not find root ${listRootName} for file ${filePath}`,
-				)
-			}
-		}
-		for (const [rootName, rootMap] of tree.entries()) {
-			const realRoot = realRoots.get(rootName)
-			if (!realRoot) {
-				console.error(`💥 Could not find root ${rootName}`)
-				continue
-			}
-			const realRootContents = fs.readdirSync(realRoot)
-			for (const rootContent of realRootContents) {
-				if (!rootMap.has(rootContent)) {
-					const pathForRemoval = path.join(realRoot, rootContent)
-					console.log(`🧹 globalFlush: removing directory ${pathForRemoval}`)
-					fs.rmSync(pathForRemoval, { recursive: true })
-				}
-			}
-			for (const [listName, list] of rootMap.entries()) {
-				const realList = path.join(realRoot, listName)
-				const realListContents = fs.readdirSync(realList)
-				for (const realListContent of realListContents) {
-					const contentTrimmed = realListContent
-						.replace(`.input.json`, ``)
-						.replace(`.output.json`, ``)
-						.replace(`stream.txt`, ``)
-					if (!list.has(contentTrimmed)) {
-						const pathForRemoval = path.join(realList, realListContent)
-						console.log(`🧹 globalFlush: removing file ${pathForRemoval}`)
-						fs.rmSync(pathForRemoval)
-					}
-				}
-			}
-		}
-		fs.rmSync(storage.rootDir, { recursive: true })
-	}
-}
-
-function startsWith<T extends string>(
-	prefix: T,
-	str: string,
-): str is `${T}${string}` {
-	return str.startsWith(prefix)
 }
